% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{assortativity_local_par}
\alias{assortativity_local_par}
\title{Calculate local assortativity in parallel}
\usage{
assortativity_local_par(graph, val, alpha = 0.2)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Calculate local assortativity in parallel
}
