% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{compile_assortativity_data}
\alias{compile_assortativity_data}
\title{Compile assortativities list into df}
\usage{
compile_assortativity_data(assortativities_list)
}
\arguments{
\item{assortativities_list}{list outputed by generate_data function}
}
\value{
Tibble containig distribution parameters of each simulation
}
\description{
Outputs parameters of local assortativity distributions in the tibble format.
Includes moments of diistributions up to kurtosis, quantiles and correlation between distributions.
}
