% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{create_b_model}
\alias{create_b_model}
\title{Create neural network with estimation of uncertainty through dropout}
\usage{
create_b_model(
  combined_frame,
  output = "surv",
  transformation = NULL,
  model = NULL
)
}
\arguments{
\item{output}{predicted value, one of "surv", "pseudo_eras", "expan"}

\item{transformation}{mathematical transformation applied to predicted value (function)}
}
\description{
Create neural network with estimation of uncertainty through dropout
}
