% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{evaluate_b_model}
\alias{evaluate_b_model}
\title{Evaluate model performance}
\usage{
evaluate_b_model(
  combined_frame,
  c_model_res,
  output = "surv",
  transformation = NULL
)
}
\description{
Evaluate model performance
}
