% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_data_fast}
\alias{generate_data_fast}
\title{Quickly perform simulations and measure assortativities for a given graph}
\usage{
generate_data_fast(
  graph,
  simulation_parameters,
  alpha = 0.7,
  n_try = 3,
  with_push = F
)
}
\arguments{
\item{graph}{input undirected graph with weight edge attribute}

\item{simulation_parameters}{survival and expansion parameters as 2 column dataframe}

\item{alpha}{damping factor for assortativity measurement}

\item{n_try}{number of repeats for each entry in simulation parameters}

\item{with_push}{implement repulsing force in simulation}
}
\description{
Quickly perform simulations and measure assortativities for a given graph
}
