% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_Bhattacharyya_dist}
\alias{get_Bhattacharyya_dist}
\title{Calculate Bhattacharyya distance to reference distribution}
\usage{
get_Bhattacharyya_dist(ref_dist, measure_dists, which_ass = 2)
}
\arguments{
\item{ref_dist}{reference distribution as datapoints}

\item{measure_dists}{list of distributions}

\item{which_ass}{which assortativity to read from data}
}
\value{
distance value for each measure distribution
}
\description{
Calculate Bhattacharyya distance to reference distribution
}
