% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_complete_results_bayes}
\alias{get_complete_results_bayes}
\title{Predict regeneration for a given graph}
\usage{
get_complete_results_bayes(
  graph,
  survival_intervals = c(0.5, 1),
  rel_expansion_intervals = c(0, 1),
  with_push = T,
  alpha = c(0.4, 0.7),
  dsize_surv = 4000,
  dsize_expan = 3000
)
}
\arguments{
\item{graph}{undirected graph with weight edge attribute and colour node attribute}

\item{survival_intervals}{range of values of survival fraction for simulation. If it is a single value, simulation and prediction of survival fraction is skipped and only expansion fraction is predicted}

\item{rel_expansion_intervals}{range of values of expansion fraction for simulation}

\item{with_push}{implement repulsing force in simulation?}

\item{alpha}{damping factor/factors}
}
\value{
returns list containing predictions, intermediate data, models and their evaluation
}
\description{
Predict regeneration for a given graph
}
