% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_predictions}
\alias{get_predictions}
\title{Get prediction for new data}
\usage{
get_predictions(assortativities_original, simulation_data, model_bayes)
}
\arguments{
\item{assortativities_original}{data compiled using compile_assortativity_data function}

\item{simulation_data}{data used for model creation. Output of preprocess_simulation function}

\item{model_bayes}{generated model}
}
\value{
model predictions for new data
}
\description{
Get prediction for new data
}
