% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{layer_concrete_dropout}
\alias{layer_concrete_dropout}
\title{Title}
\usage{
layer_concrete_dropout(
  object,
  layer,
  weight_regularizer = 1e-06,
  dropout_regularizer = 1e-05,
  init_min = 0.1,
  init_max = 0.1,
  is_mc_dropout = TRUE,
  name = NULL,
  trainable = TRUE
)
}
\arguments{
\item{object}{object}

\item{layer}{layer}

\item{weight_regularizer}{weight_regularizer}

\item{dropout_regularizer}{dropout_regularizer}

\item{init_min}{init_min}

\item{init_max}{init_max}

\item{is_mc_dropout}{is_mc_dropout}

\item{name}{name}

\item{trainable}{bool}
}
\description{
Title
}
