% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_expansion_of_fraction_cheap}
\alias{simulate_expansion_of_fraction_cheap}
\title{Simulate regeneration with colors}
\usage{
simulate_expansion_of_fraction_cheap(
  graph,
  survivor_fraction = 0.3,
  expanding_fraction = 0.1,
  cols,
  with_push = F
)
}
\arguments{
\item{graph}{undirected graph with edge weight parameter}

\item{survivor_fraction}{fraction of cells that survive}

\item{expanding_fraction}{fraction of cells among survivors that multiply and drive regeneration}

\item{cols}{vectors of cell colors. Suggested are of length equal to graph's node count}

\item{with_push}{whenever to apply repulsing force to multiplying cells}
}
\value{
list containing vectors of surviving cells, multiplying cells, cells after expansion without non_multiplying cells and finally with non_multiplying cells.
The results of this function reference colors, with the exception of surviving cells vector. It will also reference cells before
}
\description{
Simulate regeneration with colors
}
