% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{test_functionality}
\alias{test_functionality}
\title{Test package functionality}
\usage{
test_functionality()
}
\value{
result of get_complete_results
}
\description{
Run core function, get_complete_results on a test graph included in package. prediction quality is reduced, by limiting number simulations.
}
