% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{test_graph}
\alias{test_graph}
\title{Small sample graph}
\format{
Igraph graph. Important fr functionality are vertex attribute "colour", and edge attribute "weights"
}
\usage{
test_graph
}
\description{
Small graph with 355 vertices, fragment of vessels of mouse 21 days after irradiation. Used for testing package installation
}
\keyword{datasets}
